% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tximport.R
\name{tximport}
\alias{tximport}
\title{Import transcript-level abundances and estimated counts for gene-level analysis packages}
\usage{
tximport(files, type = c("kallisto", "salmon", "rsem", "cufflinks"),
  txIn = TRUE, txOut = FALSE, countsFromAbundance = c("no", "scaledTPM",
  "lengthScaledTPM"), gene2tx = NULL, geneIdCol, txIdCol, abundanceCol,
  countsCol, lengthCol, importer = function(x) read.table(x, header = TRUE),
  collatedFiles)
}
\arguments{
\item{files}{a character vector of filenames for the transcript-level abundances}

\item{type}{character, the type of software used to generate the abundances,
which will be used to autofill the arguments below (geneIdCol, etc.)}

\item{txIn}{logical, whether the incoming files are transcript level (default TRUE)}

\item{txOut}{logical, whether the function should just output transcript-level (default FALSE)}

\item{countsFromAbundance}{character, either "no" (default), "scaledTPM", or "lengthScaledTPM",
for whether to generate estimated counts using abundance estimates scaled up to library size
(scaledTPM) or additionally scaled using the average transcript length over samples and
the library size (lengthScaledTPM). if using scaledTPM or lengthScaledTPM,
then the counts are no longer correlated with average transcript length,
and so the length offset matrix should not be used.}

\item{gene2tx}{a two-column data.frame linking gene id (column 1) and transcript id (column 2).
the column names are not relevant, but this column order must be used.
this is necessary for software which does not provide such information in the file
(kallisto and Salmon)}

\item{geneIdCol}{name of column with gene id. if missing, the gene2tx argument can be used}

\item{txIdCol}{name of column with tx id}

\item{abundanceCol}{name of column with abundances (e.g. TPM or FPKM)}

\item{countsCol}{name of column with estimated counts}

\item{lengthCol}{name of column with feature length information}

\item{importer}{a function used to read in the files}

\item{collatedFiles}{a character vector of filenames for software which provides
abundances and counts in matrix form (e.g. Cufflinks). The files should be, in order,
abundances, counts, and a third file with length information}
}
\value{
a simple list with matrices: abundance, counts, length.
A final element 'countsFromAbundance' carries through
the character argument used in the tximport call.
The length matrix contains the average transcript length for each
gene which can be used as an offset for gene-level analysis.
}
\description{
Import transcript-level abundances and estimated counts for gene-level analysis packages
}

