% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tximport.R
\name{tximport}
\alias{summarizeToGene}
\alias{tximport}
\title{Import transcript-level abundances and estimated counts for gene-level analysis packages}
\usage{
tximport(files, type = c("none", "kallisto", "salmon", "sailfish", "rsem"),
  txIn = TRUE, txOut = FALSE, countsFromAbundance = c("no", "scaledTPM",
  "lengthScaledTPM"), tx2gene = NULL, reader = read.delim, geneIdCol,
  txIdCol, abundanceCol, countsCol, lengthCol, importer, collatedFiles,
  ignoreTxVersion = FALSE)

summarizeToGene(txi, tx2gene, ignoreTxVersion = FALSE,
  countsFromAbundance = c("no", "scaledTPM", "lengthScaledTPM"))
}
\arguments{
\item{files}{a character vector of filenames for the transcript-level abundances}

\item{type}{character, the type of software used to generate the abundances.
Options are "kallisto", "salmon", "sailfish", "rsem".
This argument is used to autofill the arguments below (geneIdCol, etc.)
"none" means that the user will specify these columns.}

\item{txIn}{logical, whether the incoming files are transcript level (default TRUE)}

\item{txOut}{logical, whether the function should just output transcript-level (default FALSE)}

\item{countsFromAbundance}{character, either "no" (default), "scaledTPM", or "lengthScaledTPM",
for whether to generate estimated counts using abundance estimates scaled up to library size
(scaledTPM) or additionally scaled using the average transcript length over samples and
the library size (lengthScaledTPM). if using scaledTPM or lengthScaledTPM, 
then the counts are no longer correlated with average transcript length,
and so the length offset matrix should not be used.}

\item{tx2gene}{a two-column data.frame linking transcript id (column 1) to gene id (column 2).
the column names are not relevant, but this column order must be used. 
this argument is required for gene-level summarization for methods
that provides transcript-level estimates only
(kallisto, Salmon, Sailfish)}

\item{reader}{a function to replace read.delim in the pre-set importer functions,
for example substituting read_tsv from the readr package will substantially 
speed up tximport}

\item{geneIdCol}{name of column with gene id. if missing, the gene2tx argument can be used}

\item{txIdCol}{name of column with tx id}

\item{abundanceCol}{name of column with abundances (e.g. TPM or FPKM)}

\item{countsCol}{name of column with estimated counts}

\item{lengthCol}{name of column with feature length information}

\item{importer}{a function used to read in the files}

\item{collatedFiles}{a character vector of filenames for software which provides
abundances and counts in matrix form (e.g. Cufflinks). The files should be, in order,
abundances, counts, and a third file with length information}

\item{ignoreTxVersion}{logical, whether to split the tx id on the '.' character
to remove version information, for easier matching with the tx id in gene2tx
(default FALSE)}

\item{txi}{list of matrices of trancript-level abundances, counts, and
lengths produced by \code{tximport}, only used by \code{summarizeToGene}}
}
\value{
a simple list with matrices: abundance, counts, length.
A final element 'countsFromAbundance' carries through
the character argument used in the tximport call.
The length matrix contains the average transcript length for each
gene which can be used as an offset for gene-level analysis.
Note: tximport does not import bootstrap estimates from kallisto, Salmon, or Sailfish.
}
\description{
\code{tximport} helps to import transcript-level estimates from various
external software and can summarize abundances, counts, and transcript lengths
to the gene-level (default) or output transcript-level (see \code{txOut} argument).
}
\details{
\code{summarizeToGene} can be used to summarize a list of transcript-level
matrices produced by \code{tximport} to the gene level. This is equivalent
to \code{tximport} with \code{txOut=FALSE}.

The last known supported versions of the external quantifiers are:
kallisto 0.42.4, Salmon 0.6.0, Sailfish 0.9.0, RSEM 1.2.11.
}
\section{Functions}{
\itemize{
\item \code{summarizeToGene}: Summarize tx-level matrices to gene-level
}}
\examples{

# load data for demonstrating tximport
# note that the vignette shows more examples
# including how to read in files quickly using the readr package

library(tximportData)
dir <- system.file("extdata", package="tximportData")
samples <- read.table(file.path(dir,"samples.txt"), header=TRUE)
files <- file.path(dir,"salmon", samples$run, "quant.sf")
names(files) <- paste0("sample",1:6)

# tx2gene links transcript IDs to gene IDs for summarization
tx2gene <- read.csv(file.path(dir, "tx2gene.csv"))

txi <- tximport(files, type="salmon", tx2gene=tx2gene)

}

